Classes Coordinates and Direction moved from package day1 to location
