+++ src/day1/Day1.java
@@ -1,5 +1,7 @@
 package src.day1;
 
+import src.location.Coordinates;
+import src.location.Direction;
 import src.reading.PuzzleReading;
 
 import java.io.IOException;
